Adjusted RuntimeLayer and EditorLayer to reflect the state of the editor
The RunetimeLayer will now only be active when the game is 'running' and will not respond to editor events. Likewise the editor layer will only be active while editing and will allow the user to edit the scene. Both layers will load the scene when they are attached
